requirements.txt
# This file is to be used with pip to install required python packages
# Usage: pip install -r requirements.txt

PyYAML==5.4.1

# REST/HTTP
fastapi==0.65.3
uvicorn==0.14.0
aiofiles==0.7.0
fastapi-restful==0.4.3

# Persistence
tinydb==4.4.0
#mysql-connector-python==8.0.22
FastAPI-SQLAlchemy==0.2.1

# Retry
#tenacity==6.2.0

beautifulsoup4==4.10.0
requests==2.27.1